refactor: Move handleClaim to vesting-plans
